fix: #api api/v1/auth/register
