+++ src/services/user.service.ts
@@ -3,9 +3,24 @@ import { uuid } from 'uuidv4';
 import { User } from '../entities/user.entity';
 import { db } from '../server';
 
-export async function getUserByEmail(email: string) {
+export async function getUserByEmail(email: string): Promise<User | null> {
   return await db.getRepository(User).findOneBy({
     email: email,
+  });
+}
+
+export async function createToken(id: number, email: string): Promise<string> {
+  return await sign({ id, email }, process.env.ACCESS_TOKEN_SECRET as Secret, {
+    expiresIn: '1h',
+  });
+}
+
+export async function createRefreshToken(
+  id: number,
+  email: string
+): Promise<string> {
+  return await sign({ id, email }, process.env.ACCESS_TOKEN_SECRET as Secret, {
+    expiresIn: '24h',
   });
 }
 
@@ -17,20 +32,9 @@ export async function createUser(email: string, password: string) {
   });
   const user = await db.getRepository(User).save(newUser);
 
-  const access_token = await sign(
-    { id: user.id, email: user.email },
-    process.env.ACCESS_TOKEN_SECRET as Secret,
-    {
-      expiresIn: '1h',
-    }
-  );
-  const refresh_token = await sign(
-    { id: user.id, email: user.email },
-    process.env.ACCESS_TOKEN_SECRET as Secret,
-    {
-      expiresIn: '24h',
-    }
-  );
+  const access_token = await createToken(user.id, user.email);
+  const refresh_token = await createRefreshToken(user.id, user.email);
+
   return {
     access_token,
     refresh_token,
